+++ bin/commands/state.py
@@ -56,6 +56,7 @@ def state(**kwargs):
     :keyword bool show_empty: show empty sections
     :keyword list order: order to print sections in
     :keyword bool clear: clear terminal before printing
+    :keyword bool page: page output if too long
     """
 
     if not directories.is_git_repository():
@@ -149,7 +150,7 @@ def state(**kwargs):
     state_result = state_result[:-1] # strip the extra trailing newline
     state_lines = len(state_result.splitlines())
     terminal_lines = literal_eval(subprocess.check_output(['tput', 'lines']))
-    if terminal_lines >= state_lines + 2: # one for the newline and one for the prompt
+    if not kwargs.get('page', True) or terminal_lines >= state_lines + 2: # one for the newline and one for the prompt
         if kwargs.get('clear') and sys.stdout.isatty():
             subprocess.call('clear')
         messages.info(state_result)  # TODO: breaks --no-color